% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_mean_dataset.R
\docType{data}
\name{bee_mean_dataset}
\alias{bee_mean_dataset}
\title{Body size and intertegular distance measurements of ~400 bee species.}
\format{An object of class \code{data.frame} with 535 rows and 11 columns.}
\usage{
bee_mean_dataset
}
\description{
A dataset containing the mean body size in dry weight (mg), intertegular distance (mm) and metadata for ~400 bee species.
}
\keyword{datasets}
