% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forage_dataset.R
\docType{data}
\name{forage_dataset}
\alias{forage_dataset}
\title{Bee foraging range measurements for 101 bee species.}
\format{
An object of class \code{data.frame} with 425 rows and 20 columns.
}
\usage{
forage_dataset
}
\description{
A dataset containing measurements of foraging range (m), intertegular distance (mm), sociality and metadata for 101 bee species (425 estimates).
}
\references{
\itemize{
\item Kendall et al. (2022). Sociality and body size determine the potential and realized foraging ranges of bees. \url{TBC}.#' }
}
\keyword{datasets}
