% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foraging.range.R
\name{foraging.range}
\alias{foraging.range}
\title{Uses intertegular distance, and degree of sociality, to predict foraging range for bees.}
\usage{
foraging.range(
  data,
  model.type = "social",
  random.effects = "reduced",
  measure.type = "realized"
)
}
\arguments{
\item{data}{A dataframe containing between two to four columns: bee intertegular distance (ITD) measurements in mm, species (Format: "Genus_species") and sociality (Format: "Highly Eusocial","Primitively Eusocial", "Solitary").}

\item{model.type}{A string stating which model should be used: "social", for the model that includes sociality, body size (ITD) and measurement type, or "ITD", for the reduced model that only includes body size and measurement type.}

\item{random.effects}{A sting stating if random effects are to be used using predictive purposes. Options are 1) phylogenetic effect and species effects "full", 
2) only phylogenetic effect ("phylo"), or 3) fixed effects only ("reduced"). Default is "reduced".}

\item{measure.type}{A string stating whether to return realized ("realized"), potential ("potential") estimates.}
}
\value{
A dataframe with bee foraging range (km) is returned for each bee species (row).
}
\description{
Calculates realized* foraging range from Kendall et al. (2022) using intertegular distance (ITD) values
 and degree of sociality (highly eusocial, primitively eusocial or solitary). *Potential estimates are not recommended to be used but can be obtained by selecting `measure.type = "potential"`.
}
\examples{
ITD=c(2.9,2.9,2.9)
species=c("Apis_mellifera","Osmia_cornifrons","Bombus_bifarius")
sociality=c("Highly Eusocial","Solitary","Primitively Eusocial")
example.data <- data.frame(ITD,species,sociality)
foraging.range(example.data,random.effects = "full",measure.type = "realized")

}
\references{
\itemize{
\item Kendall et al. (2022) Sociality and body size determine the potential and realized foraging ranges of bees \emph{TBC} \url{TBC}.
}
}
