% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headwidthsize.R
\name{headwidthsize}
\alias{headwidthsize}
\title{Converts pollinator (Diptera, Hymenoptera and Lepidoptera) head width (mm) to body weight (mg).}
\usage{
headwidthsize(HW, Eq = "H97DB")
}
\arguments{
\item{HW}{A vector of head width measurments (mm).}

\item{Eq}{a vector of predictive allometries for insect taxa from Hodar (1997). 
Options implemented are: H97DB (Brachycera), H97DN (Nematocera), H97HA (Hymenoptera), H97LH (Heterocera) and H97LR (Rhopalocera).}
}
\value{
A dataframe with pollinator body size (mg) is returned for each specimen from selected equation.
}
\description{
Calculates dry body weight (mg) using head width (mm) from Hodar (1997).
}
\examples{
headwidthsize(HW=c(10,5,2), Eq = c("H97DB","H97DB","H97DB")
}
\references{
Kendall, Bartomeus... Rader (20XX) Pollinator size and its consequences: Predictive allometry for pollinating insects 
#ADD Hodar 1997
}
