% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hov_mean_dataset.R
\docType{data}
\name{hov_mean_dataset}
\alias{hov_mean_dataset}
\title{Body size and intertegular distance measurements of ~100 hoverfly species.}
\format{An object of class \code{data.frame} with 148 rows and 10 columns.}
\usage{
hov_mean_dataset
}
\description{
A dataset containing the mean body size in dry weight (mg), intertegular distance (mm) and metadata for 100 hoverfly species.
}
\keyword{datasets}
