% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITweigth.R
\name{ITweigth}
\alias{ITweigth}
\title{Converts IT measures to body mass for bees.}
\usage{
ITweigth(IT)
}
\arguments{
\item{IT}{A vector of bee intertegular spans (IT) measurments in cm.}

\item{Family}{A vector of bee families. Accepted names are c("", "",...)}
}
\value{
A dataframe with bee body masses (gr) is returned for each bees species.
}
\description{
Calculates body mass form ..., using intertegular distance values (IT).
}
\examples{
ITweigth(c(10,5,2))

}
\references{
Cane, J. (1987) Estimation of bee size using intertegular span (Apoidea). Journal of the Kansas Entomological Society, 60, 145-147.
}
