% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Melin_et_al_2019.R
\docType{data}
\name{Melin_et_al_2019}
\alias{Melin_et_al_2019}
\title{Bee tongue length measurements for 11 Melittidae bee species.}
\format{An object of class \code{data.frame} with 179 rows and 7 columns.}
\usage{
Melin_et_al_2019
}
\description{
A dataset containing glossa (mm), prementum (mm) and proboscis (mm) and taxonomic information for 179 measured bee specimens.
}
\details{
@references Melin et al. (2019) The allometry of proboscis length in Melittidae (Hymenoptera: Apoidae) and an estimate of their foraging distance using museum specimens. PloS one, 14(6), e0217839. <doi:10.1371/journal.pone.0217839>
}
\keyword{datasets}
