% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forage.R
\name{SizeForage}
\alias{SizeForage}
\title{Converts body weight to measures of foraging distance for bees.}
\usage{
SizeForage(BW, Type = "GreenleafAll")
}
\arguments{
\item{BW}{A vector of either bee intertegular spans (IT) measurments in cm, head width values in mm or dry weight values (mg).}

\item{Type}{The type of foraging distance desired. Options are "GrAll", GrMhd ("Maximum homing distance"),
GrThd ("Typical homing distance"), GrMfd ("Maximum feeder training distance"), GrMcd("Maximum communication distance"),"Osmia", "MeliMR" , "MeliFT" or "MeliAll". See details in Greenleaf et al. (2007), Guédot et al. (2009) and van Nieuwstadt and Iraheta (1996) and}
}
\value{
A dataframe with bee and foraging distance (Km) 
is returned for each bees species.
mhd: Maximum homing distance
thd: Typical homing distance
mfd: Maximum feeder training distance
mcd: Maximum communication distance
MeliMR: Meliponini mark-recapture 95% return distance
MeliFT: Meliponini feeder training 95% foraging distance
}
\description{
Calculates foraging distance from Greenleaf et al. (2007) (using intertegular distance values (IT)),
 van Nieuwstadt and Iraheta (1996) using head width values for Meliponini (stingless bees) and 
 Guédot et al. (2009) using dry weight values (mg) for Osmia species.
}
\examples{
SizeForage(c(10,5,2))
SizeForage(c(10,5,2), type = "MeliMR") 

}
\references{
Greenleaf, S.S., Williams, N.M., Winfree, R. & Kremen, C. (2007) Bee foraging ranges and their relationship to body size. Oecologia, 153, 589-596.
}
