% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tongue.R
\name{tongue}
\alias{tongue}
\title{Converts IT measures tongue length for bees.}
\usage{
tongue(IT, family, mouthpart = "all")
}
\arguments{
\item{IT}{A vector of bee intertegular spans (IT) measurments in cm.}

\item{family}{a vector of bee families. Only implemented 5 out of the extant 7 families: 
"Andrenidae", "Apidae", "Colletidae", "Halictidae", "Megachilidae".}

\item{mouthpart}{The mouth part you are interested in. Options are "all", glossa", "prementum" and "tongue" (i.e. gloss + prementum)}
}
\value{
A dataframe with bee tongue length (mm) is returned for each bees species.
}
\description{
Calculates tongue length (mm) from Cariveau et al. (2015) using intertegular distance values (IT).
}
\examples{
ITtongue(c(10,5,2), family = c("Andrenidae", "Apidae", "Colletidae"))
ITtongue(c(10,5,2), family = c("Andrenidae", "Apidae", "Colletidae"), mouthpart = "tongue")
}
\references{
Cariveau, Nayak, Bartomeus, Zientek, Ascher, Winfree (2015) The allometry of bee tongue length an its uses in ecology and evolution
}
