% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodysize.R
\name{bodysize}
\alias{bodysize}
\title{Converts the intertegular distance (ITD) and co-variates to body size in dry weight (mg) for bees or hoverflies.}
\usage{
bodysize(x, taxa, type)
}
\arguments{
\item{x}{A data frame with columns containing ITD values and Sex ('Male' or 'Female'). 
Optional: Family/Subfamily, Region (Only NorthAmerica, SouthAmerica, Australasia and Europe) and Species ("Genus_species")}

\item{taxa}{A vector specifying insect taxa of interest, can be either "bee" for bee models and "hov" for hoverfly models}

\item{type}{A vector specifying model type to be used: for bees this can be either "taxo" for taxonomic models, "phy" for phylogenetic model or "IT" for ITD-only model. In hoverflies: it can either be "h1" for Weight ~ IT + Sex, "h2" for Weight ~ IT * Subfamily + Sex or IT for ITD-only model.}
}
\value{
The original dataframe (x) is returned along with four additional columns: body size (dry weight (mg)), S.E. and 95% confidence intervals.
}
\description{
Calculates body size in dry weight (mg) from Kendall et al. (2018) using ITD (and co-variate) values.
}
\details{
For bees, type option 'taxo' requires ITD, 
sex and taxonomic family.   Option 'phylo' 
 is only recommended for species included in model formulation. 
 For hoverflies, type 'h1' requires ITD and Sex for each specimen 
 and type 'h2' requires ITD, Sex and subfamily. Type "IT" for both 
 taxa only requires ITD values (Optional: region and species, if included). 
 If specimens are from included regions (see above) or  species we recommend 
 including these as additional columns.
}
\examples{
example=cbind.data.frame(IT=1:2,
Sex=c("Female","Male"), 
Family=c("Apidae","Andrenidae"),
Region=c("NorthAmerica","Europe"),
Species=c("Ceratina_dupla","Andrena_flavipes"))
bodysize(x=example,taxa="bee",type="taxo")

}
\references{
Kendall, Rader, Gagic... Bartomeus (2018) Pollinator size and its consequences: Predictive allometry for pollinating insects
}
