% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollimetry_dataset.R
\docType{data}
\name{pollimetry_dataset}
\alias{pollimetry_dataset}
\title{Body size and intertegular distance measurements of ~4500 pollinating insect specimens.}
\format{An object of class \code{data.frame} with 4434 rows and 25 columns.}
\usage{
pollimetry_dataset
}
\description{
A dataset containing the body size in dry weight (mg), intertegular distance (mm), body length (mg) and metadata for 4438 measured bee and hoverfly specimens.
}
\details{
@references Kendall et al. (2018)  Pollinator size and its consequences: Predictive allometry for pollinating insects. <doi:10.1101/397604>
}
\keyword{datasets}
