% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollimetry_dataset.R
\docType{data}
\name{pollimetry_dataset}
\alias{pollimetry_dataset}
\title{Body size and intertegular distance measurements of ~4500 pollinating insect specimens.}
\format{An object of class \code{data.frame} with 4434 rows and 25 columns.}
\usage{
pollimetry_dataset
}
\description{
A dataset containing the body size in dry weight (mg), intertegular distance (mm), body length (mg) and metadata for 4438 measured bee and hoverfly specimens.
}
\details{
@references Kendall et al. (2019) Pollinator size and its consequences: Robust estimates of body size in pollinating insects. Ecology and Evolution, 9(4), 1702-1714. <doi:10.1002/ece3.4835>
}
\keyword{datasets}
