% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tonguelength.R
\name{tonguelength}
\alias{tonguelength}
\title{Converts ITD (cm) to tongue length for bees.}
\usage{
tonguelength(IT, family, mouthpart = "all")
}
\arguments{
\item{IT}{A vector of bee intertegular distance (IT) measurments in cm.}

\item{family}{a vector of bee families. Only implemented 5 out of the extant 7 families: 
"Andrenidae", "Apidae", "Colletidae", "Halictidae", "Megachilidae".}

\item{mouthpart}{The mouth part you are interested in. Options are "all", glossa", "prementum" and "tongue" (i.e. gloss + prementum)}
}
\value{
A dataframe with bee tongue length (mm) is returned for each bees species.
}
\description{
Calculates tongue length (mm) from Cariveau et al. (2015) using intertegular distance (ITD) values (cm)..
}
\examples{
tonguelength(IT=c(10,5,2), family = c("Andrenidae", "Apidae", "Colletidae"))
tonguelength(IT=c(10,5,2), family = c("Andrenidae", "Apidae", "Colletidae"), mouthpart = "tongue")

}
\references{
Cariveau et al. (2016) The allometry of bee tongue length an its uses in ecology and evolution. PLoS ONE 11(3): e0151482 <doi:10.1371/journal.pone.0151482>
}
