% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tonguelength.R
\name{tonguelength}
\alias{tonguelength}
\title{Converts ITD (mm) to tongue length for bees.}
\usage{
tonguelength(x, mouthpart = "all")
}
\arguments{
\item{x}{A dataframe with the following two columns: bee intertegular distance (IT) measurements in mm and family, a vector of bee families. Only six out of the extant seven families are implemented: 
"Andrenidae", "Apidae", "Colletidae", "Halictidae", "Melittidae" and "Megachilidae".}

\item{mouthpart}{The mouthpart you are interested in. Options are: "all", "glossa", "prementum" and "tongue" (i.e. gloss + prementum)}
}
\value{
A dataframe with bee tongue length (mm) and 95% confidence intervals are returned for each bee species bound to your original dataframe.
}
\description{
Calculates tongue length (mm) from Cariveau et al. (2015) and Melin et al. (2019) using intertegular distance (ITD) values (mm).
}
\examples{
example=cbind.data.frame(IT=c(1.3,2.3),
                         Family=c("Andrenidae","Apidae"))
tonguelength(example,mouthpart="all")

}
\references{
\itemize{
\item Kendall et al. (2019) Pollinator size and its consequences: Robust estimates of body size in pollinating insects. \emph{Ecology and Evolution}, 9(4), 1702-1714. \doi{10.1002/ece3.4835}.
\item Cariveau et al. (2016) The allometry of bee tongue length an its uses in ecology and evolution. \emph{PloS one}, 11(3): e0151482. \doi{10.1371/journal.pone.0151482}.
\item Melin et al. (2019) The allometry of proboscis length in Melittidae (Hymenoptera: Apoidae) and an estimate of their foraging distance using museum specimens. \emph{PloS one}, 14(6), e0217839. \doi{10.1371/journal.pone.0217839}.
}
}
